fix music modules integration
