+++ .idea/gradle.xml
@@ -21,7 +21,6 @@
             <option value="$PROJECT_DIR$/favorite/ui" />
             <option value="$PROJECT_DIR$/music" />
             <option value="$PROJECT_DIR$/music/data" />
-            <option value="$PROJECT_DIR$/music/detail" />
             <option value="$PROJECT_DIR$/music/domain" />
             <option value="$PROJECT_DIR$/music/ui" />
             <option value="$PROJECT_DIR$/playlist" />
