+++ .idea/misc.xml
@@ -21,6 +21,12 @@
         <entry key="music/detail/src/main/res/layout/activity_music_detail.xml" value="0.1" />
         <entry key="music/detail/src/main/res/layout/frag_music_detail.xml" value="0.1" />
         <entry key="music/detail/src/main/res/menu/menu_music_detail.xml" value="0.1" />
+        <entry key="music/ui/src/main/res/drawable/check_favorite.xml" value="0.1325" />
+        <entry key="music/ui/src/main/res/drawable/ic_star_border_white_24dp.xml" value="0.1325" />
+        <entry key="music/ui/src/main/res/drawable/ic_star_white_24dp.xml" value="0.1325" />
+        <entry key="music/ui/src/main/res/layout/activity_music_detail.xml" value="0.1" />
+        <entry key="music/ui/src/main/res/layout/frag_music_detail.xml" value="0.1" />
+        <entry key="music/ui/src/main/res/menu/menu_music_detail.xml" value="0.1" />
         <entry key="playlist/create/src/main/res/layout/frag_add_playlist.xml" value="0.1" />
         <entry key="playlist/create/src/main/res/menu/menu_playlist_detail.xml" value="0.1" />
         <entry key="playlist/detail/src/main/res/layout/frag_playlist_detail.xml" value="0.1" />
@@ -29,6 +35,8 @@
         <entry key="playlist/src/main/res/menu/menu_context_playlist.xml" value="0.1037037037037037" />
         <entry key="playlist/ui/src/main/res/layout/frag_add_playlist.xml" value="0.1828774062816616" />
         <entry key="playlist/ui/src/main/res/layout/frag_playlist_detail.xml" value="0.1828774062816616" />
+        <entry key="playlist/ui/src/main/res/layout/playlist_fragment.xml" value="0.1" />
+        <entry key="playlist/ui/src/main/res/menu/menu_context_playlist.xml" value="0.1" />
         <entry key="playlist/ui/src/main/res/menu/menu_playlist_detail.xml" value="0.37025641025641026" />
         <entry key="search/src/main/res/layout/frag_music_search.xml" value="0.1" />
         <entry key="search/src/main/res/layout/item_list_footer.xml" value="0.1" />
