--- music/detail/src/main/AndroidManifest.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.jeanpaulo.musiclibrary.music.detail">
-
-    <application>
-        <activity
-            android:exported="true"
-            android:name=".presentation.MusicDetailActivity"
-            android:screenOrientation="portrait"
-            android:windowSoftInputMode="adjustResize"
-            android:label="@string/title_activity_music_detail" />
-    </application>
-</manifest>
