--- music/detail/src/main/java/com/jeanpaulo/musiclibrary/music/detail/presentation/model/MusicDetailUIModel.kt
@@ -1,13 +1,0 @@
-package com.jeanpaulo.musiclibrary.music.detail.presentation.model
-
-data class MusicDetailUIModel(
-    val id: Long,
-    val name: String,
-    val artwork: String,
-    val previewURL: String?,
-    val artist: String,
-    val album: String,
-) {
-    fun hasPreview() = previewURL != null
-    fun getPreview(): String = previewURL ?: ""
-}
