+++ music/ui/src/main/AndroidManifest.xml
@@ -1,5 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.jeanpaulo.musiclibrary.music.ui">
-
+    <application>
+        <activity
+            android:exported="true"
+            android:name="com.jeanpaulo.musiclibrary.music.ui.MusicDetailActivity"
+            android:screenOrientation="portrait"
+            android:windowSoftInputMode="adjustResize"
+            android:label="@string/title_activity_music_detail" />
+    </application>
 </manifest>
