+++ music/ui/src/main/java/com/jeanpaulo/musiclibrary/music/ui/di/MusicDetailQualifier.kt
@@ -1,0 +1,7 @@
+package com.jeanpaulo.musiclibrary.music.ui.di
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class SimpleMusicUI
