+++ music/ui/src/main/java/com/jeanpaulo/musiclibrary/music/ui/model/MusicDetailUIModel.kt
@@ -1,0 +1,13 @@
+package com.jeanpaulo.musiclibrary.music.ui.model
+
+data class MusicDetailUIModel(
+    val id: Long,
+    val name: String,
+    val artwork: String,
+    val previewURL: String?,
+    val artist: String,
+    val album: String,
+) {
+    fun hasPreview() = previewURL != null
+    fun getPreview(): String = previewURL ?: ""
+}
