+++ music/ui/src/main/res/drawable/check_favorite.xml
