+++ music/ui/src/main/res/drawable/ic_star_border_white_24dp.xml
