+++ music/ui/src/main/res/drawable/ic_star_white_24dp.xml
