+++ music/ui/src/main/res/layout/frag_music_detail.xml
