+++ music/ui/src/main/res/menu/menu_music_detail.xml
